+++ bedrock/tests/test_smart_account_safe_owner_swap.rs
@@ -1,25 +1,18 @@
 use std::sync::Arc;
 
 use alloy::{
-    network::Ethereum,
-    primitives::{keccak256, U256},
-    providers::{ext::AnvilApi, Provider, ProviderBuilder},
+    primitives::U256,
+    providers::{ext::AnvilApi, ProviderBuilder},
     signers::local::PrivateKeySigner,
     sol,
-    sol_types::SolValue,
 };
 use bedrock::{
-    primitives::{
-        http_client::{set_http_client, AuthenticatedHttpClient, HttpError, HttpHeader, HttpMethod},
-    },
+    primitives::http_client::set_http_client,
     smart_account::{SafeSmartAccount, ENTRYPOINT_4337},
-    transaction::foreign::UnparsedUserOperation,
 };
-use serde::Serialize;
-use serde_json::json;
 
 mod common;
-use common::{deploy_safe, setup_anvil, IEntryPoint, PackedUserOperation};
+use common::{deploy_safe, setup_anvil, AnvilBackedHttpClient, IEntryPoint};
 
 sol! {
     /// Safe owner management interface
@@ -28,184 +21,6 @@ sol! {
         function getOwners() external view returns (address[] memory);
         function isOwner(address owner) external view returns (bool);
         function swapOwner(address prevOwner, address oldOwner, address newOwner) external;
-    }
-}
-
-// ------------------ Mock HTTP client that intercepts sponsor and executes on Anvil ------------------
-#[derive(Clone)]
-struct AnvilBackedHttpClient<P>
-where
-    P: Provider<Ethereum> + Clone + Send + Sync + 'static,
-{
-    provider: P,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct SponsorUserOperationResponseLite<'a> {
-    paymaster: &'a str,
-    paymaster_data: &'a str,
-    pre_verification_gas: String,
-    verification_gas_limit: String,
-    call_gas_limit: String,
-    paymaster_verification_gas_limit: String,
-    paymaster_post_op_gas_limit: String,
-    max_priority_fee_per_gas: String,
-    max_fee_per_gas: String,
-}
-
-#[async_trait::async_trait]
-impl<P> AuthenticatedHttpClient for AnvilBackedHttpClient<P>
-where
-    P: Provider<Ethereum> + Clone + Send + Sync + 'static,
-{
-    async fn fetch_from_app_backend(
-        &self,
-        _url: String,
-        method: HttpMethod,
-        _headers: Vec<HttpHeader>,
-        body: Option<Vec<u8>>,
-    ) -> Result<Vec<u8>, HttpError> {
-        if method != HttpMethod::Post {
-            return Err(HttpError::Generic {
-                message: "unsupported method".into(),
-            });
-        }
-
-        let body = body.ok_or(HttpError::Generic {
-            message: "missing body".into(),
-        })?;
-
-        let root: serde_json::Value =
-            serde_json::from_slice(&body).map_err(|_| HttpError::Generic {
-                message: "invalid json".into(),
-            })?;
-
-        let method =
-            root.get("method")
-                .and_then(|m| m.as_str())
-                .ok_or(HttpError::Generic {
-                    message: "invalid json".into(),
-                })?;
-        let id = root.get("id").cloned().unwrap_or(serde_json::Value::Null);
-        let params = root
-            .get("params")
-            .cloned()
-            .unwrap_or(serde_json::Value::Null);
-
-        match method {
-            // Intercept sponsor request and return minimal gas values with no paymaster
-            "wa_sponsorUserOperation" => {
-                let result = SponsorUserOperationResponseLite {
-                    paymaster: "0x0000000000000000000000000000000000000000",
-                    paymaster_data: "0x",
-                    pre_verification_gas: "0x20000".into(),
-                    verification_gas_limit: "0x20000".into(),
-                    call_gas_limit: "0x20000".into(),
-                    paymaster_verification_gas_limit: "0x0".into(),
-                    paymaster_post_op_gas_limit: "0x0".into(),
-                    max_priority_fee_per_gas: "0x3B9ACA00".into(), // 1 gwei
-                    max_fee_per_gas: "0x3B9ACA00".into(),          // 1 gwei
-                };
-
-                let response = json!({
-                    "jsonrpc": "2.0",
-                    "id": id,
-                    "result": result,
-                });
-
-                Ok(serde_json::to_vec(&response).unwrap())
-            }
-            // Execute the user operation on Anvil
-            "eth_sendUserOperation" => {
-                let params = params.as_array().ok_or(HttpError::Generic {
-                    message: "invalid params".into(),
-                })?;
-                let user_op_val = params.first().ok_or(HttpError::Generic {
-                    message: "missing userOp param".into(),
-                })?;
-                let _entry_point_str = params.get(1).and_then(|v| v.as_str()).ok_or(
-                    HttpError::Generic {
-                        message: "missing entryPoint param".into(),
-                    },
-                )?;
-
-                // Build UnparsedUserOperation from JSON (which uses hex strings), then convert
-                let obj = user_op_val.as_object().ok_or(HttpError::Generic {
-                    message: "userOp param must be an object".into(),
-                })?;
-
-                let get_opt = |k: &str| -> Option<String> {
-                    obj.get(k).and_then(|v| v.as_str()).map(|s| s.to_string())
-                };
-                let get_or_zero = |k: &str| -> String {
-                    get_opt(k).unwrap_or_else(|| "0x0".to_string())
-                };
-                let get_required = |k: &str| -> Result<String, HttpError> {
-                    get_opt(k).ok_or(HttpError::Generic {
-                        message: format!("missing or invalid {k}"),
-                    })
-                };
-
-                let unparsed = UnparsedUserOperation {
-                    sender: get_required("sender")?,
-                    nonce: get_required("nonce")?,
-                    call_data: get_required("callData")?,
-                    call_gas_limit: get_or_zero("callGasLimit"),
-                    verification_gas_limit: get_or_zero("verificationGasLimit"),
-                    pre_verification_gas: get_or_zero("preVerificationGas"),
-                    max_fee_per_gas: get_or_zero("maxFeePerGas"),
-                    max_priority_fee_per_gas: get_or_zero("maxPriorityFeePerGas"),
-                    paymaster: get_opt("paymaster"),
-                    paymaster_verification_gas_limit: get_or_zero(
-                        "paymasterVerificationGasLimit",
-                    ),
-                    paymaster_post_op_gas_limit: get_or_zero("paymasterPostOpGasLimit"),
-                    paymaster_data: get_opt("paymasterData"),
-                    signature: get_required("signature")?,
-                    factory: get_opt("factory"),
-                    factory_data: get_opt("factoryData"),
-                };
-
-                let parsed_op: bedrock::smart_account::UserOperation =
-                    unparsed.try_into().map_err(|e| HttpError::Generic {
-                        message: format!("invalid userOp: {e}"),
-                    })?;
-
-                // Execute on Anvil via EntryPoint
-                let entry_point_contract = IEntryPoint::new(*ENTRYPOINT_4337, &self.provider);
-                let packed_op = PackedUserOperation::try_from(&parsed_op)
-                    .map_err(|e| HttpError::Generic {
-                        message: format!("failed to pack user operation: {e}"),
-                    })?;
-                let handle_ops = entry_point_contract
-                    .handleOps(vec![packed_op], parsed_op.sender)
-                    .gas(5_000_000)
-                    .send()
-                    .await
-                    .map_err(|e| HttpError::Generic {
-                        message: format!("failed to execute user operation: {e}"),
-                    })?;
-
-                let _receipt = handle_ops.get_receipt().await.map_err(|e| HttpError::Generic {
-                    message: format!("failed to get receipt: {e}"),
-                })?;
-
-                // Create a user operation hash
-                let user_op_hash = keccak256(parsed_op.abi_encode());
-
-                let response = json!({
-                    "jsonrpc": "2.0",
-                    "id": id,
-                    "result": format!("0x{}", hex::encode(user_op_hash)),
-                });
-
-                Ok(serde_json::to_vec(&response).unwrap())
-            }
-            _ => Err(HttpError::Generic {
-                message: format!("unsupported method: {method}"),
-            }),
-        }
     }
 }
 
@@ -276,16 +91,13 @@ async fn test_safe_owner_swap_e2e() -> anyhow::Result<()> {
     // Execute the owner swap using tx_swap_safe_owner
     println!("→ Executing tx_swap_safe_owner to swap owners...");
     let tx_hash = safe_account
-        .tx_swap_safe_owner(
-            &initial_owner.to_string(),
-            &new_owner.to_string(),
-        )
+        .tx_swap_safe_owner(&initial_owner.to_string(), &new_owner.to_string())
         .await?;
 
-    println!("✓ Executed owner swap transaction: {}", tx_hash.to_hex_string());
-    
-    // Wait a bit for the transaction to be processed
-    tokio::time::sleep(tokio::time::Duration::from_secs(2)).await;
+    println!(
+        "✓ Executed owner swap transaction: {}",
+        tx_hash.to_hex_string()
+    );
 
     // Verify the owner swap was successful
     let final_owners = safe_contract.getOwners().call().await?;
